Initial Metalsmith
This is a static site, so we're going to use a static site
generator to generate it. This PR kicks off that vision.

We decided to use Metalsmith for some reason, though there
are concerns that it isn't an established enough tool. It's
quite possible we will rip out Metalsmith and replace it with
something more robust like Jekyll.