Remove source of flashing
There are more in depth fixes but in the short term this flash is a liability for anybody prone to seizures.

This disables our constantly updating reading time, but that's OK since it was causing harm.

We'll make a longer term fix that restores the functionality at a later date.

Related to Issue #99