chore(eslint): disable some annoying rules
react/jsx-props-no-spreading: We spread props, a lot. It very often improperly
adds attributes to DOM elements when it shouldn't, so I get why this rule exists.
For now the strategy is to refactor spreading to NOT pass those attributes to
DOM (see console messages for warnings), but not completely remove prop spreading.

react/require-default-props: Default props are being removed from function
components in the future, which throws warnings. So if we don't use default props
anymore, then stop warning us when we don't have them.