refactor(PersonModal): port typescript (see note)
this is a proof of concept for components that ingest graphql data.

graphql queries assume all fields are nullable by default. typescript would then
demand we handle all null values, which sucks and will make code less readable
unless we can guarantee the fields are never null in the graphql schema.
see here
https://www.gatsbyjs.com/docs/how-to/local-development/graphql-typegen/#non-nullable-types

in the case of people, we can guarantee certain fields, e.g. "html" is always present,
so is "frontmatter". fields like "bluesky" is definitely null-ish but i'm not sure if
it's actually null or just, say, empty string