Port JavaScript to TypeScript, Phase 1 (#134)
* chore(typescript): add basic tsconfig.json

* refactor(404): port typescript

* chore(ts): add ui path to config

* refactor(settings): port typescript

* refactor(mixins): port most to typescript

* refactor(themes): port typescript

* refactor(brandmarks): port typescript

* refactor(actionbars): port typescript

* refactor(Footer): port typescript

also brandmark props should be optional bc they have default values

* refactor(assets): add type definitions for imported assets

* refactor(animations): port typescript

* refactor(partials/ofLayout): port typescript

* refactor(contexts): port typescript

* refactor(partials): port most of them to typescript

and add types for react-helmet

* refactor(assets): port remainder to typescript

* refactor(Layout): port typescript + convert to functional component

* refactor(setSpace): port to typescript

not done! see ts errors in setSpace.ts

* refactor(Gallery): port typescript

* refactor(Icon): port typescript

* refactor(Tabs): port typescript

* refactor(Copy): port typescript

* refactor(Tile): port typescript

not done; there are types to pass thru because <Tile> is polymorphic and
can be rendered as <a> with its associated props/attributes

* refactor(Modal): port typescript

* chore(gatsby): auto-generate gatsby query types

* refactor(PersonModal): port typescript (see note)

this is a proof of concept for components that ingest graphql data.

graphql queries assume all fields are nullable by default. typescript would then
demand we handle all null values, which sucks and will make code less readable
unless we can guarantee the fields are never null in the graphql schema.
see here
https://www.gatsbyjs.com/docs/how-to/local-development/graphql-typegen/#non-nullable-types

in the case of people, we can guarantee certain fields, e.g. "html" is always present,
so is "frontmatter". fields like "bluesky" is definitely null-ish but i'm not sure if
it's actually null or just, say, empty string

* refactor(Tile): polymorphic component typing proof of concept

i am kind of muddling through this attempt at making <Tile> work as a
polymorphic component because in some cases it renders out an <a> instead
of a <div>. right now it doesn't really cover the condition where if
`as="a"` is passed as a prop, then the html attributes of <a> are allowed

* refactor(actions): port typescript (see notes)

the problem here is that when <Action> is used, it can be a <Button> or a <Link>
or it can be an obfuscated version of those (see `withObfucscation`). What
it becomes is dependent on the combination of props and its values that passed to
<Action>. i'm either not good enough at typescript to know what to do here or
maybe we would have to revisit this approach????

* refactor(Link): do not need to type theme parameter on styled components